Created ServuceBase and implemented it on ActivityService Cleaned up ActivityService after moving some methods to ServiceBase
